inference/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
google-cloud-storage>=2.16.0
pydantic>=2.7.0
